requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
elastic-transport==8.17.0
elasticsearch==8.17.1
exceptiongroup==1.2.2
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
injector==0.22.0
ollama==0.4.7
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
ruff==0.9.6
sniffio==1.3.1
starlette==0.45.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
